Add a method to extract all orders from user history
While web_orders_gogdata() is able to acquire the most recent operations
in a user's history, occasionally we wish to go through its entirety.
This can be used to calculate the full amount already spent, determine
when a game was purchased or do more elaborate analysis such as how a
user changes behaviour or preferences since they created their GOG
account.

The new method was added with what are considered reasonable default
arguments.
At the same time, a new URL was added, supporting all the search
arguments for this query.